code cleanup,tweaks and new renderer
the new renderer is much more accurate but due to some positioning
issues it has i have made active for only when backgrounds are rendered.
the new renderer has an async load at the end(not because i've optimized
it its actually a limitation). in order to use it you have to call the
background command once to load and then with a slight delay call it
again and it will actually return the image.